# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InvestorStatus = typeof InvestorStatus[keyof typeof InvestorStatus];


export const InvestorStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Investor {
  id: number;
  fullName: string;
  email: string;
  status: InvestorStatus;
  createdAt: string;
}

export type InvestorAdminStatus = typeof InvestorAdminStatus[keyof typeof InvestorAdminStatus];


export const InvestorAdminStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface InvestorAdmin {
  id: number;
  fullName: string;
  email: string;
  status: InvestorAdminStatus;
  createdAt: string;
  shares: string;
  avgCost: string;
}

export interface InvestorInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
}

export interface InvestorCount {
  count: number;
}

export type InvestorStatusInputStatus = typeof InvestorStatusInputStatus[keyof typeof InvestorStatusInputStatus];


export const InvestorStatusInputStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface InvestorStatusInput {
  status: InvestorStatusInputStatus;
}

export interface CreditInput {
  /** @minimum 0 */
  shares: number;
  /** @minimum 0 */
  pricePerShare: number;
}

export interface SignInInput {
  email: string;
}

export type SignInResultStatus = typeof SignInResultStatus[keyof typeof SignInResultStatus];


export const SignInResultStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SignInResult {
  status: SignInResultStatus;
  fullName: string;
  email: string;
}

export interface Holdings {
  investorId: number;
  shares: string;
  avgCost: string;
  cashBalance: string;
  updatedAt: string;
}

export interface SiteConfig {
  sellingEnabled: boolean;
}

export interface SiteConfigUpdate {
  sellingEnabled: boolean;
}

export interface BuyInput {
  email: string;
  /** @minimum 0.01 */
  amountUsd: number;
}

export interface SellInput {
  email: string;
  /** @minimum 0.0001 */
  shares: number;
}

export type CreateDepositInputMethod = typeof CreateDepositInputMethod[keyof typeof CreateDepositInputMethod];


export const CreateDepositInputMethod = {
  card: 'card',
  crypto: 'crypto',
} as const;

export type CreateDepositInputCoin = typeof CreateDepositInputCoin[keyof typeof CreateDepositInputCoin];


export const CreateDepositInputCoin = {
  BTC: 'BTC',
  ETH: 'ETH',
  DOGE: 'DOGE',
} as const;

export interface CreateDepositInput {
  email: string;
  /** @minimum 1 */
  amount: number;
  method: CreateDepositInputMethod;
  coin?: CreateDepositInputCoin;
}

export type DepositRecordMethod = typeof DepositRecordMethod[keyof typeof DepositRecordMethod];


export const DepositRecordMethod = {
  card: 'card',
  crypto: 'crypto',
} as const;

export type DepositRecordStatus = typeof DepositRecordStatus[keyof typeof DepositRecordStatus];


export const DepositRecordStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface DepositRecord {
  id: number;
  investorId: number;
  email: string;
  amount: string;
  method: DepositRecordMethod;
  /** @nullable */
  coin: string | null;
  status: DepositRecordStatus;
  createdAt: string;
}

export type DepositAdminMethod = typeof DepositAdminMethod[keyof typeof DepositAdminMethod];


export const DepositAdminMethod = {
  card: 'card',
  crypto: 'crypto',
} as const;

export type DepositAdminStatus = typeof DepositAdminStatus[keyof typeof DepositAdminStatus];


export const DepositAdminStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface DepositAdmin {
  id: number;
  investorId: number;
  fullName: string;
  email: string;
  amount: string;
  method: DepositAdminMethod;
  /** @nullable */
  coin: string | null;
  status: DepositAdminStatus;
  createdAt: string;
}

export interface DepositAddress {
  coin: string;
  address: string;
  updatedAt: string;
}

export interface UpdateDepositAddressInput {
  /** @minLength 10 */
  address: string;
}

export interface StockQuote {
  symbol: string;
  price: number;
  change: number;
  changePct: number;
  volume: string;
  marketCap: string;
  bid: number;
  ask: number;
  dayLow: number;
  dayHigh: number;
  week52High: number;
  week52Low: number;
  /** @nullable */
  peRatio: string | null;
  avgVolume: string;
  sharesOut: string;
}

export interface StockDataPoint {
  time: string;
  price: number;
}

export interface StockHistory {
  period: string;
  data: StockDataPoint[];
}

export interface ErrorResponse {
  error: string;
}

export type GetStockHistoryParams = {
period?: GetStockHistoryPeriod;
};

export type GetStockHistoryPeriod = typeof GetStockHistoryPeriod[keyof typeof GetStockHistoryPeriod];


export const GetStockHistoryPeriod = {
  '1D': '1D',
  '1W': '1W',
  '1M': '1M',
  '3M': '3M',
  '1Y': '1Y',
  '5Y': '5Y',
} as const;

export type GetHoldingsParams = {
email: string;
};

export type ListDepositsParams = {
email: string;
};