# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deposits:
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List deposits for the current user (by email query param)
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deposit history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositRecord"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/investors:
    get:
      operationId: listAdminInvestors
      tags: [admin]
      summary: List all investors (admin only)
      responses:
        "200":
          description: List of investors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestorAdmin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/investors/{id}/status:
    patch:
      operationId: updateInvestorStatus
      tags: [admin]
      summary: Update investor status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestorStatusInput"
      responses:
        "200":
          description: Updated investor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorAdmin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/investors/{id}/credit:
    post:
      operationId: creditInvestor
      tags: [admin]
      summary: Manually credit shares to an investor (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditInput"
      responses:
        "200":
          description: Updated holdings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holdings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deposits:
    get:
      operationId: listAdminDeposits
      tags: [admin]
      summary: List all deposits (admin only)
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositAdmin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deposit-addresses:
    get:
      operationId: adminGetDepositAddresses
      tags: [admin]
      summary: Get all deposit addresses (admin only)
      responses:
        "200":
          description: Deposit addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositAddress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/site-config:
    patch:
      operationId: updateSiteConfig
      tags: [admin]
      summary: Update platform configuration, e.g. enable/disable investor selling (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteConfigUpdate"
      responses:
        "200":
          description: Updated site configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deposit-addresses/{coin}:
    put:
      operationId: updateDepositAddress
      tags: [admin]
      summary: Update a deposit address (admin only)
      parameters:
        - name: coin
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepositAddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositAddress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    InvestorAdmin:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
        shares:
          type: string
        avgCost:
          type: string
      required:
        - id
        - fullName
        - email
        - status
        - createdAt
        - shares
        - avgCost
...
    InvestorStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
      required:
        - status
    CreditInput:
      type: object
      properties:
        shares:
          type: number
          minimum: 0
        pricePerShare:
          type: number
          minimum: 0
      required:
        - shares
        - pricePerShare
...
    Holdings:
      type: object
      properties:
        investorId:
          type: integer
        shares:
          type: string
        avgCost:
          type: string
        cashBalance:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - investorId
        - shares
        - avgCost
        - cashBalance
        - updatedAt
    SiteConfig:
      type: object
      properties:
        sellingEnabled:
          type: boolean
      required:
        - sellingEnabled
    SiteConfigUpdate:
      type: object
      properties:
        sellingEnabled:
          type: boolean
      required:
        - sellingEnabled
...
    CreateDepositInput:
      type: object
      properties:
        email:
          type: string
        amount:
          type: number
          minimum: 1
        method:
          type: string
          enum: [card, crypto]
        coin:
          type: string
          enum: [BTC, ETH, DOGE]
      required:
        - email
        - amount
        - method
    DepositRecord:
      type: object
      properties:
        id:
          type: integer
        investorId:
          type: integer
        email:
          type: string
        amount:
          type: string
        method:
          type: string
          enum: [card, crypto]
        coin:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - investorId
        - email
        - amount
        - method
        - coin
        - status
        - createdAt
    DepositAdmin:
      type: object
      properties:
        id:
          type: integer
        investorId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        amount:
          type: string
        method:
          type: string
          enum: [card, crypto]
        coin:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - investorId
        - fullName
        - email
        - amount
        - method
        - coin
        - status
        - createdAt
    DepositAddress:
      type: object
      properties:
        coin:
          type: string
        address:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - coin
        - address
        - updatedAt
    UpdateDepositAddressInput:
      type: object
      properties:
        address:
          type: string
          minLength: 10
      required:
        - address
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error